Move contents of cta.go into generate.go
